Major netflix work
I believe there is still an issue remaining with the netflix callback.
I need to publish to azure and then  work on the callback that way.

Some controller work was done to handle different states of the logged
in/out and netflix linked/unlinked.

Split out the models to be more organized.